fixed loading issue and browser console logs
