+++ .gitignore
@@ -34,6 +34,7 @@ yarn-error.log*
 .env.production.local
 !packages/database/.env
 !apps/web/.env
+.env.example
 
 # Prisma generated files
 packages/database/zod
