+++ apps/web/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/summary/components/shareEmbedTabs/WebpageTab.tsx
@@ -17,6 +17,7 @@ export default function WebpageTab({ surveyUrl }) {
     <div className="flex h-full grow flex-col gap-5">
       <div className="flex justify-between">
         <div className=""></div>
+
         <Button
           variant="darkCTA"
           title="Embed survey in your website"
